[CDAP-12230] Add RunRecordCorrectorService
