+++ cdap-app-fabric/src/main/java/co/cask/cdap/internal/app/services/ProgramLifecycleService.java
@@ -33,22 +33,16 @@ import co.cask.cdap.common.ConflictException;
 import co.cask.cdap.common.NotFoundException;
 import co.cask.cdap.common.ProgramNotFoundException;
 import co.cask.cdap.common.app.RunIds;
-import co.cask.cdap.common.conf.CConfiguration;
-import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.common.io.CaseInsensitiveEnumTypeAdapterFactory;
 import co.cask.cdap.common.security.AuthEnforce;
-import co.cask.cdap.common.service.Retries;
-import co.cask.cdap.common.service.RetryStrategies;
 import co.cask.cdap.config.PreferencesStore;
 import co.cask.cdap.internal.app.ApplicationSpecificationAdapter;
-import co.cask.cdap.internal.app.runtime.AbstractListener;
 import co.cask.cdap.internal.app.runtime.BasicArguments;
 import co.cask.cdap.internal.app.runtime.ProgramOptionConstants;
 import co.cask.cdap.internal.app.runtime.SimpleProgramOptions;
 import co.cask.cdap.internal.app.runtime.schedule.ProgramSchedule;
 import co.cask.cdap.internal.app.runtime.schedule.ProgramScheduleStatus;
 import co.cask.cdap.internal.app.store.RunRecordMeta;
-import co.cask.cdap.proto.BasicThrowable;
 import co.cask.cdap.proto.Id;
 import co.cask.cdap.proto.NamespaceMeta;
 import co.cask.cdap.proto.ProgramRecord;
@@ -70,12 +64,7 @@ import co.cask.cdap.security.spi.authentication.AuthenticationContext;
 import co.cask.cdap.security.spi.authorization.AuthorizationEnforcer;
 import co.cask.cdap.security.spi.authorization.UnauthorizedException;
 import co.cask.cdap.store.NamespaceStore;
-import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Supplier;
-import com.google.common.base.Throwables;
-import com.google.common.collect.Collections2;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Sets;
 import com.google.common.util.concurrent.AbstractIdleService;
 import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
@@ -84,7 +73,6 @@ import com.google.gson.GsonBuilder;
 import com.google.inject.Inject;
 import org.apache.twill.api.RunId;
 import org.apache.twill.api.logging.LogEntry;
-import org.apache.twill.common.Threads;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -93,14 +81,10 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.EnumSet;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
 import javax.annotation.Nullable;
 
 /**
@@ -114,10 +98,8 @@ public class ProgramLifecycleService extends AbstractIdleService {
     .registerTypeAdapterFactory(new CaseInsensitiveEnumTypeAdapterFactory())
     .create();
 
-  private final ScheduledExecutorService scheduledExecutorService;
   private final Store store;
   private final ProgramRuntimeService runtimeService;
-  private final CConfiguration cConf;
   private final NamespaceStore nsStore;
   private final PropertiesResolver propertiesResolver;
   private final PreferencesStore preferencesStore;
@@ -127,15 +109,13 @@ public class ProgramLifecycleService extends AbstractIdleService {
 
   @Inject
   ProgramLifecycleService(Store store, NamespaceStore nsStore, ProgramRuntimeService runtimeService,
-                          CConfiguration cConf, PropertiesResolver propertiesResolver,
+                          PropertiesResolver propertiesResolver,
                           PreferencesStore preferencesStore, AuthorizationEnforcer authorizationEnforcer,
                           AuthenticationContext authenticationContext, Scheduler scheduler) {
     this.store = store;
     this.nsStore = nsStore;
     this.runtimeService = runtimeService;
     this.propertiesResolver = propertiesResolver;
-    this.scheduledExecutorService = Executors.newScheduledThreadPool(1);
-    this.cConf = cConf;
     this.preferencesStore = preferencesStore;
     this.authorizationEnforcer = authorizationEnforcer;
     this.authenticationContext = authenticationContext;
@@ -146,27 +126,12 @@ public class ProgramLifecycleService extends AbstractIdleService {
   protected void startUp() throws Exception {
     LOG.info("Starting ProgramLifecycleService");
 
-    long interval = cConf.getLong(Constants.AppFabric.PROGRAM_RUNID_CORRECTOR_INTERVAL_SECONDS);
-    if (interval <= 0) {
-      LOG.debug("Invalid run id corrector interval {}. Setting it to 180 seconds.", interval);
-      interval = 180L;
-    }
-    scheduledExecutorService.scheduleWithFixedDelay(new RunRecordsCorrectorRunnable(this),
-                                                    2L, interval, TimeUnit.SECONDS);
+
   }
 
   @Override
   protected void shutDown() throws Exception {
     LOG.info("Shutting down ProgramLifecycleService");
-
-    scheduledExecutorService.shutdown();
-    try {
-      if (!scheduledExecutorService.awaitTermination(5, TimeUnit.SECONDS)) {
-        scheduledExecutorService.shutdownNow();
-      }
-    } catch (InterruptedException ie) {
-      Thread.currentThread().interrupt();
-    }
   }
 
   /**
@@ -763,145 +728,6 @@ public class ProgramLifecycleService extends AbstractIdleService {
   }
 
   /**
-   * Fix all the possible inconsistent states for RunRecords that shows it is in RUNNING state but actually not
-   * via check to {@link ProgramRuntimeService}.
-   */
-  private void validateAndCorrectRunningRunRecords() {
-    Set<String> processedInvalidRunRecordIds = Sets.newHashSet();
-
-    // Lets update the running programs run records
-    for (ProgramType programType : ProgramType.values()) {
-      validateAndCorrectRunningRunRecords(programType, processedInvalidRunRecordIds);
-    }
-
-    if (!processedInvalidRunRecordIds.isEmpty()) {
-      LOG.info("Corrected {} of run records with RUNNING status but no actual program running.",
-               processedInvalidRunRecordIds.size());
-    }
-  }
-
-  /**
-   * Fix all the possible inconsistent states for RunRecords that shows it is in RUNNING state but actually not
-   * via check to {@link ProgramRuntimeService} for a type of CDAP program.
-   *
-   * @param programType The type of program the run records need to validate and update.
-   * @param processedInvalidRunRecordIds the {@link Set} of processed invalid run record ids.
-   */
-  @VisibleForTesting
-  void validateAndCorrectRunningRunRecords(final ProgramType programType,
-                                           final Set<String> processedInvalidRunRecordIds) {
-    final Map<RunId, RuntimeInfo> runIdToRuntimeInfo = runtimeService.list(programType);
-
-    LOG.trace("Start getting run records not actually running ...");
-    Collection<RunRecordMeta> notActuallyRunning = store.getRuns(ProgramRunStatus.RUNNING,
-                                                           new com.google.common.base.Predicate<RunRecordMeta>() {
-      @Override
-      public boolean apply(RunRecordMeta input) {
-        String runId = input.getPid();
-        // Check if it is not actually running.
-        return !runIdToRuntimeInfo.containsKey(RunIds.fromString(runId));
-      }
-    }).values();
-    LOG.trace("End getting {} run records not actually running.", notActuallyRunning.size());
-
-    final Map<String, ProgramId> runIdToProgramId = new HashMap<>();
-
-    LOG.trace("Start getting invalid run records  ...");
-    Collection<RunRecordMeta> invalidRunRecords =
-      Collections2.filter(notActuallyRunning, new com.google.common.base.Predicate<RunRecordMeta>() {
-        @Override
-        public boolean apply(RunRecordMeta input) {
-          String runId = input.getPid();
-          // check for program Id for the run record, if null then it is invalid program type.
-          ProgramId targetProgramId = retrieveProgramIdForRunRecord(programType, runId);
-
-          // Check if run id is for the right program type
-          if (targetProgramId != null) {
-            runIdToProgramId.put(runId, targetProgramId);
-            return true;
-          } else {
-            return false;
-          }
-        }
-      });
-
-    // don't correct run records for programs running inside a workflow
-    // for instance, a MapReduce running in a Workflow will not be contained in the runtime info in this class
-    invalidRunRecords = Collections2.filter(invalidRunRecords, new com.google.common.base.Predicate<RunRecordMeta>() {
-      @Override
-      public boolean apply(RunRecordMeta invalidRunRecordMeta) {
-        boolean shouldCorrect = shouldCorrectForWorkflowChildren(invalidRunRecordMeta, processedInvalidRunRecordIds);
-        if (!shouldCorrect) {
-          LOG.trace("Will not correct invalid run record {} since it's parent workflow still running.",
-                    invalidRunRecordMeta);
-          return false;
-        }
-        return true;
-      }
-    });
-
-    LOG.trace("End getting invalid run records.");
-
-    if (!invalidRunRecords.isEmpty()) {
-      LOG.warn("Found {} RunRecords with RUNNING status and the program not actually running for program type {}",
-               invalidRunRecords.size(), programType.getPrettyName());
-    } else {
-      LOG.trace("No RunRecords found with RUNNING status and the program not actually running for program type {}",
-                programType.getPrettyName());
-    }
-
-    // Now lets correct the invalid RunRecords
-    for (RunRecordMeta invalidRunRecordMeta : invalidRunRecords) {
-      String runId = invalidRunRecordMeta.getPid();
-      ProgramId targetProgramId = runIdToProgramId.get(runId);
-
-      boolean updated = store.compareAndSetStatus(targetProgramId, runId, ProgramController.State.ALIVE.getRunStatus(),
-                                                  ProgramController.State.ERROR.getRunStatus());
-      if (updated) {
-        LOG.warn("Fixed RunRecord {} for program {} with RUNNING status because the program was not " +
-                   "actually running",
-                 runId, targetProgramId);
-
-        processedInvalidRunRecordIds.add(runId);
-      }
-    }
-  }
-
-  /**
-   * Helper method to check if the run record is a child program of a Workflow
-   *
-   * @param runRecordMeta The target {@link RunRecordMeta} to check
-   * @param processedInvalidRunRecordIds the {@link Set} of processed invalid run record ids.
-   * @return {@code true} of we should check and {@code false} otherwise
-   */
-  private boolean shouldCorrectForWorkflowChildren(RunRecordMeta runRecordMeta,
-                                                   Set<String> processedInvalidRunRecordIds) {
-    // check if it is part of workflow because it may not have actual runtime info
-    if (runRecordMeta.getProperties() != null && runRecordMeta.getProperties().get("workflowrunid") != null) {
-
-      // Get the parent Workflow info
-      String workflowRunId = runRecordMeta.getProperties().get("workflowrunid");
-      if (!processedInvalidRunRecordIds.contains(workflowRunId)) {
-        // If the parent workflow has not been processed, then check if it still valid
-        ProgramId workflowProgramId = retrieveProgramIdForRunRecord(ProgramType.WORKFLOW, workflowRunId);
-        if (workflowProgramId != null) {
-          // lets see if the parent workflow run records state is still running
-          RunRecordMeta wfRunRecord = store.getRun(workflowProgramId, workflowRunId);
-          RuntimeInfo wfRuntimeInfo = runtimeService.lookup(workflowProgramId, RunIds.fromString(workflowRunId));
-
-          // Check of the parent workflow run record exists and it is running and runtime info said it is still there
-          // then do not update it
-          if (wfRunRecord != null && wfRunRecord.getStatus() == ProgramRunStatus.RUNNING && wfRuntimeInfo != null) {
-            return false;
-          }
-        }
-      }
-    }
-
-    return true;
-  }
-
-  /**
    * Helper method to get {@link ProgramId} for a RunRecord for type of program
    *
    * @param programType Type of program to search
@@ -909,7 +735,7 @@ public class ProgramLifecycleService extends AbstractIdleService {
    * @return the program id of the run record or {@code null} if does not exist.
    */
   @Nullable
-  private ProgramId retrieveProgramIdForRunRecord(ProgramType programType, String runId) {
+  protected ProgramId retrieveProgramIdForRunRecord(ProgramType programType, String runId) {
 
     // Get list of namespaces (borrow logic from AbstractAppFabricHttpHandler#listPrograms)
     List<NamespaceMeta> namespaceMetas = nsStore.list();
@@ -1072,35 +898,5 @@ public class ProgramLifecycleService extends AbstractIdleService {
       throw new BadRequestException("Update log levels at runtime is only supported in distributed mode");
     }
     return ((LogLevelUpdater) programController);
-  }
-
-  /**
-   * Helper class to run in separate thread to validate the invalid running run records
-   */
-  private static class RunRecordsCorrectorRunnable implements Runnable {
-    private static final Logger LOG = LoggerFactory.getLogger(RunRecordsCorrectorRunnable.class);
-
-    private final ProgramLifecycleService programLifecycleService;
-
-    RunRecordsCorrectorRunnable(ProgramLifecycleService programLifecycleService) {
-      this.programLifecycleService = programLifecycleService;
-    }
-
-    @Override
-    public void run() {
-      try {
-        RunRecordsCorrectorRunnable.LOG.trace("Start correcting invalid run records ...");
-
-        // Lets update the running programs run records
-        programLifecycleService.validateAndCorrectRunningRunRecords();
-
-        RunRecordsCorrectorRunnable.LOG.trace("End correcting invalid run records.");
-      } catch (Throwable t) {
-        // Ignore any exception thrown since this behaves like daemon thread.
-        //noinspection ThrowableResultOfMethodCallIgnored
-        LOG.warn("Unable to complete correcting run records: {}", Throwables.getRootCause(t).getMessage());
-        LOG.debug("Exception thrown when running run id cleaner.", t);
-      }
-    }
   }
 }
